requirements.txt
discord
requests
python-dateutil